--- assets/Colors.js
@@ -1,20 +1,0 @@
-const Colors = {
-  // Green
-  lightestGreen: '#C9E3D9',
-  lighterGreen: '#48AC69',
-  primaryGreen: '#008550',
-  darkerGreen: '#004E14',
-  // Text
-  activeText: 'rgba(0, 0, 0, 0.8)',
-  secondaryText: '#999999',
-  // Other
-  error: '#CB4D4D',
-  base: '#BDBDBD',
-  light: '#DEDEDE',
-  lighter: '#EBEBEB',
-  lightest: '#FFFFFF',
-  black: '#000000',
-  trainingMode: '#FFC300',
-};
-
-export default Colors;
