+++ navigation/AppNavigator.js
@@ -3,8 +3,8 @@ import { AsyncStorage, Linking, TouchableOpacity, View } from 'react-native';
 import { createAppContainer, createSwitchNavigator } from 'react-navigation';
 import { createDrawerNavigator } from 'react-navigation-drawer';
 import { createStackNavigator } from 'react-navigation-stack';
-import Colors from '../assets/Colors';
 import { Title } from '../components/BaseComponents';
+import Colors from '../constants/Colors';
 import CheckoutScreen from '../screens/CheckoutScreen';
 import ClerkLoginScreen from '../screens/ClerkLoginScreen';
 import ConfirmationScreen from '../screens/ConfirmationScreen';
