+++ styled/checkin.js
@@ -1,6 +1,6 @@
 import { Platform } from 'react-native';
 import styled from 'styled-components/native';
-import Colors from '../assets/Colors';
+import Colors from '../constants/Colors';
 
 export const CheckInContainer = styled.KeyboardAvoidingView.attrs({
   behavior: Platform.OS === 'ios' ? 'position' : null,
